// src/types/MonthlyRevenueAndExpenses.ts
export interface MonthlyRevenueAndExpenses {
  hoa: number;
  propertyTaxes: number;
  fireInsurance: number;
  floodInsurance: number;
  pmi: number;
  repairsReservePerMonthAmount: number;
  capitalExpenditure: number;
  waterSewer: number;
  garbage: number;
  gas: number;
  electricty: number;
  snowRemoval: number;
  lawnCare: number;
  propertyManagementPercentage: number;
  other: number;
}